added ability to write and respond in the topic, updated carousel on frontend, fixed userRating of movie in search input

+++ src/backend/API/Models/Responses/MovieSearchResponse.cs
@@ -5,6 +5,6 @@ public class MovieSearchResponse
     public Guid Id { get; set; }
     public string Title { get; set; }
     public int Year { get; set; }
-    public int UserRating { get; set; }
+    public double UserRating { get; set; }
     public List<ActorSearchResponse> Actors { get; set; }
 }
